chore: move main from server to root
