+++ .gitignore
@@ -10,3 +10,4 @@
 
 # aeron & other binaries
 *.jar
+**/aeron-archive
