+++ Cargo.toml
@@ -16,6 +16,14 @@ members = [
     "server", 
 ]
 
+[dependencies]
+vex-server = { workspace = true }
+vex-config = { workspace = true }
+ctrlc = { workspace = true }
+tracing-subscriber = { workspace =  true}
+tracing = { workspace = true}
+tikv-jemallocator = { workspace = true }
+
 [workspace.dependencies]
 common = { path = "./common"}
 vex-config = { path = "./vex-config" }
@@ -36,6 +44,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tracing = "0.1"
 tracing-subscriber = "0.1"
+ctrlc = "3.4"
 parking_lot = "0.12"
 tikv-jemallocator = "0.5"
