+++ server/Cargo.toml
@@ -17,11 +17,9 @@ vex-networking = { workspace = true }
 parking_lot = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
-tracing-appender = "0.2"
 hashbrown = { workspace = true }
 disruptor = { workspace = true }
 thiserror = { workspace = true }
-ctrlc = "3.4"
 tikv-jemallocator = { workspace = true }
 
 [dev-dependencies]
